Update example_elastic_complex_freesurface.py
Added .to(device) for source_amplitudes_y because by default the deepwave.wavelets.ricker(freq, nt, dt, peak_time).reshape(1, 1, -1)
it returns a CPU tensor by default. We have to pass to the device we are working with. 

Also on line 22, can you please fix the name? I wasn't able to find the example_elastic_complex_freesurface.png in any of the 4 code examples in the documentation. Can you please fix it so that we are referencing an actual .jpg or .png file we created in the examples? Thanks!